-preparing for integration of custom bidding strategies -multiple smaller changes in the code regarding imports -removed functions from misc to utils -added new csv file for defining bidding strategies -defaults to simple if not given -edited market config to add product type -bids are performed accoridng to the product_type
