+++ assume/markets/__init__.py
@@ -1,1 +1,2 @@
-from .base_market import MarketRole
+from assume.markets.base_market import MarketRole
+from assume.markets.clearing_algorithms import pay_as_clear, pay_as_bid
