+++ assume/strategies/__init__.py
@@ -1,3 +1,3 @@
-from .base_strategy import BaseStrategy
-from .naive_strategies import NaiveStrategyMarkUp, NaiveStrategyNoMarkUp
-from .rl_strategies import RLStrategy
+from assume.strategies.base_strategy import BaseStrategy
+from assume.strategies.naive_strategies import NaiveStrategy
+from assume.strategies.rl_strategies import RLStrategy
