+++ assume/strategies/base_strategy.py
@@ -1,8 +1,3 @@
-import warnings
-from abc import ABC, abstractmethod
-from typing import Any, Dict, Generator, List, Optional, Union
-
-
 class BaseStrategy:
     """A base class for a bidding strategy.
 
@@ -18,5 +13,5 @@ class BaseStrategy:
     def __init__(self):
         super().__init__()
 
-    def calculate_bids(self, operational_window):
+    def calculate_bids(self, product, operational_window):
         raise NotImplementedError()
