+++ examples/inputs/example_01/bidding_strategies.csv
@@ -1,0 +1,6 @@
+name,energy,capacity
+KKW ISAR 2,simple,simple
+NEURATH F,simple,simple
+GKM 9,simple,simple
+IRSCHING 5,simple,simple
+demand_DE,simple,simple
