+++ tests/test_market_mechanisms.py
@@ -3,8 +3,8 @@ from datetime import datetime, timedelta
 from dateutil import rrule as rr
 from dateutil.relativedelta import relativedelta as rd
 
-from assume.common.market_mechanisms import available_clearing_strategies
-from assume.common.marketclasses import MarketConfig, MarketProduct, Order, Orderbook
+from assume.markets.clearing_algorithms import available_clearing_strategies
+from assume.common.market_objects import MarketConfig, MarketProduct, Order, Orderbook
 from assume.common.utils import get_available_products
 from assume.markets.base_market import MarketRole
